Add examples for chapter 4 w/ mods to playbooks
Add shell scripts to test examples in chapter 4.

One example clones a git repo, so we've added the git package
to the provisioning playbook.

The text refers to smilint, and ex 02 for this chapter uses smistrip.
These commands come from different packages, to the os_X group_vars
have been updated to install the necessary bits.